fix: extract focusPrevent to utils
