+++ src/script/util/util.ts
@@ -30,6 +30,7 @@ import {StorageKey} from '../storage/StorageKey';
 import {loadValue} from './StorageUtil';
 import {AuthError} from '../error/AuthError';
 import type {Conversation} from '../entity/Conversation';
+import {isTabKey} from 'Util/KeyboardUtil';
 
 export const isTemporaryClientAndNonPersistent = (persist: boolean): boolean => {
   if (persist === undefined) {
@@ -339,3 +340,24 @@ export function throttle(callback: Function, wait: number, immediate = false) {
     }
   };
 }
+
+export const preventFocusOutside = (event: KeyboardEvent, parentId: string): void => {
+  if (!isTabKey(event)) {
+    return;
+  }
+  event.preventDefault();
+  const focusableElements =
+    'button:not([disabled]), [href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), [tabindex]:not([tabindex="-1"])';
+  const parent = document.getElementById(parentId);
+  const focusableContent = [...parent.querySelectorAll(focusableElements)];
+  const focusedItemIndex = focusableContent.indexOf(document.activeElement);
+  if (event.shiftKey && focusedItemIndex === 0) {
+    (focusableContent[focusableContent.length - 1] as HTMLElement)?.focus();
+    return;
+  }
+  if (!event.shiftKey && focusedItemIndex === focusableContent.length - 1) {
+    (focusableContent[0] as HTMLElement)?.focus();
+    return;
+  }
+  (focusableContent[focusedItemIndex + 1] as HTMLElement)?.focus();
+};
